Update pages with incoming changes
